Refactor core.main into services
