+++ .env.template
@@ -3,3 +3,4 @@ DEBUG=
 SUMMARIZER_BACKEND=
 OPENAI_API_KEY=
 OPENAI_MODEL_NAME=
+TIMEOUT_MULTITHREADING=
